many many long vs int changes

+++ composeApp/src/main/kotlin/com/example/a62550_foodapp/db/dao/RecipeItemDao.kt
@@ -17,7 +17,7 @@ interface RecipeItemDao {
     suspend fun insertAll(items: List<RecipeItem>)
 
     @Query("DELETE FROM recipe_items WHERE recipe_id = :recipeId")
-    suspend fun deleteForRecipe(recipeId: Int)
+    suspend fun deleteForRecipe(recipeId: Long)
 
     // one-shot load (used when adding to shopping list)
     @Query("""
@@ -25,7 +25,7 @@ interface RecipeItemDao {
         FROM recipe_items
         WHERE recipe_id = :recipeId
     """)
-    suspend fun getItemsForRecipe(recipeId: Int): List<RecipeItem>
+    suspend fun getItemsForRecipe(recipeId: Long): List<RecipeItem>
 
     // reactive load (used in UI)
     @Query("""
@@ -33,5 +33,5 @@ interface RecipeItemDao {
         FROM recipe_items
         WHERE recipe_id = :recipeId
     """)
-    fun getItemsForRecipeFlow(recipeId: Int): Flow<List<RecipeItem>>
+    fun getItemsForRecipeFlow(recipeId: Long): Flow<List<RecipeItem>>
 }
